requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-docx==1.1.0
openai==1.13.3
python-multipart==0.0.9
markdown==3.5.2
python-dotenv==1.0.1
